Fix deprecation error in I18n
Utility\I18n's protected _bindTextDomain method was sometimes assigning an array key
to an array element that evaluated as false. This was trigging deprecation error
'Automatic conversion of false to array is deprecated'.

This modification checks whether the array element is false and explicitly converts it
to an array if it is.

closes #70